setup: Use system-wide aarch64-linux-gnu-gcc and arm-none-eabi-gcc
Migrate deprecated AOSP gcc to a separate compiler target, `aosp_gcc`,
while using the `gcc` target for a more recent and more relevant
compiler that might be installed on ones system. The script can easily
be called with `PATH=/path/to/your/arm/gcc:/and/aarch64 ./oot.sh` as
well.